+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project mostly adheres to [Semantic Versioning](https://semver.org/spec
   optionally (currently only for `ClimatoologyUserError` or `InputValidationError` results) a message (closes
   [#22](https://gitlab.heigit.org/climate-action/api-gateway/-/issues/22))
 - Restructured the repository to use multiple files separated by route
+- Cache plugin info in `get_plugin` call
 
 ### Added